+++ src/jsfuncs.ts
@@ -106,6 +106,31 @@ function jsSplice(val : HibikiVal, ...rest : any[]) {
     return newArr;
 }
 
+function jsMoveItem(val : HibikiVal, fromIndexVal : HibikiVal, toIndexVal : HibikiVal) {
+    let [arrObj, isArr] = DataCtx.asArray(val, false);
+    if (!isArr) {
+        return null;
+    }
+    let fromIndex = Math.trunc(DataCtx.valToNumber(fromIndexVal));
+    let toIndex = Math.trunc(DataCtx.valToNumber(toIndexVal));
+    if (isNaN(fromIndex)) {
+        throw new Error(sprintf("fn:moveitem 'fromindex' is NaN, type=%s", DataCtx.hibikiTypeOf(fromIndexVal)));
+    }
+    if (isNaN(toIndex)) {
+        throw new Error(sprintf("fn:moveitem 'toindex' is NaN, type=%s", DataCtx.hibikiTypeOf(toIndexVal)));
+    }
+    if (fromIndex < 0 || toIndex < 0 || fromIndex >= arrObj.length || toIndex >= arrObj.length) {
+        throw new Error(sprintf("fn:moveitem from/to indexes are out-of-bounds.  from=%d to=%d len=%d", fromIndex, toIndex, arrObj.length));
+    }
+    let newArr = [...arrObj];
+    if (fromIndex === toIndex) {
+        return newArr;
+    }
+    let elems = newArr.splice(fromIndex, 1);
+    newArr.splice(toIndex, 0, elems[0]);
+    return newArr;
+}
+
 function processSliceArgs(arg : HibikiVal, fnName : string) : [number, number] {
     if (arg == null) {
         return null;
@@ -421,7 +446,7 @@ function jsArrReverse(params : DataCtx.HibikiParamsObj, dataenv : DataEnvironmen
         console.log("WARNING fn:reverse makerefs=true, but data is not a reference");
     }
     if (!makeRefs || !(rawVal instanceof DataCtx.LValue)) {
-        return arrObj.reverse();
+        return arrObj.slice().reverse();
     }
     let rawLv = rawVal as DataCtx.LValue;
     return arrObj.map((_, idx) => rawLv.subArrayIndex(idx)).reverse();
@@ -949,6 +974,7 @@ regParamFn("reverse", jsArrReverse);
 regParamFn("every", jsArrEvery);
 regParamFn("some", jsArrSome);
 reg("concat", jsArrConcat, true);
+reg("moveitem", jsMoveItem, true);
 
 // concat