+++ changelog.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## v0.3.4
+
+* Allow parsing of YAML data if js-yaml UMD package is manually included (window.jsyaml is set)
+* Can set textDelimiters (on config and on hibiki tag as an attribute) to "none", "default", or "alt".  Selects between no inline h-text elements, {{ ... }}, and {| ... |} respectively.  The "alt" option can be used with Django templates.
+* do not fire welcome message until DOMContentLoaded event
+* new function js:moveitem(arr, from-index, to-index)
+* add new shared root ($shared) that allows data to be shared between Hibiki templates on the same page
+* export HibikiUtils module for libraries
+* (beta) added drag and drop support, handlers and properties (html5)
+* (beta) new element geometry tracking: geo.bindpath (client/offset height/width, and scrolltop)
+* (beta) new scroll tracking: scrolltop.bindpath, and scroll.handler (debounced)
+* remove ability to set window.HibikiParserOpts
+* expand foreach inside of \@children
+* bugfix: reload HibikiGlobalConfig after DOMContentLoaded event
+* bugfix: Hibiki.States was not holding the correct HibikiExtState object
+
 ## v0.3.3
 
 * HibikiGlobalConfig preRenderHook and postRenderHook (receives state and DOM node)